verson4

+++ CvarpeClient/src/com/kvrmnks/net/DownLoader.java
@@ -1,5 +1,6 @@
 package com.kvrmnks.net;
 
+import com.kvrmnks.data.MyDate;
 import com.kvrmnks.data.MyFile;
 import com.kvrmnks.data.SimpleLogListProperty;
 import com.kvrmnks.data.UserData;
@@ -16,39 +17,44 @@ import java.util.Scanner;
 public class DownLoader implements Runnable, Serializable {
     private static final int PACKAGE_SIZE = 1024;
     private long id;
-    private String pos, realPos,md5,name;
+    private String pos, realPos, md5, name;
     private boolean flag = false;
     private SimpleLogListProperty simpleLogListProperty;
     private Net server;
-    private File realFile,infoFile;
-    private long curSize,wholeSize;
+    private File realFile, infoFile;
+    private long curSize, wholeSize;
+    private NetReader serverReader;
 
-
-    public DownLoader(long id, String pos,String name, String realPos, SimpleLogListProperty simpleLogListProperty) throws IOException, NotBoundException, ClassNotFoundException, NoUserException, NoAccessException, NoFileException {
+    public DownLoader(long id, String pos, String name, String realPos, SimpleLogListProperty simpleLogListProperty) throws IOException, NotBoundException, ClassNotFoundException, NoUserException, NoAccessException, NoFileException {
         this.id = id;
         this.pos = pos;
         this.name = name;
         this.realPos = realPos;
         this.simpleLogListProperty = simpleLogListProperty;
         md5 = "";
-        server = (Net)Naming.lookup(UserData.getIp());
+        server = (Net) Naming.lookup(UserData.getServerIp());
+        serverReader = (NetReader) Naming.lookup(UserData.getReaderIp());
         buildFile();
-        MyFile myFile = server.getStructure(id,pos);
+        MyFile myFile = server.getStructure(id, pos);
         myFile = myFile.sonFile.get(name);
+        serverReader.setFile(myFile.getId(),myFile.getName());
         wholeSize = myFile.getSize();
-        simpleLogListProperty.setProcess(((double)curSize)/((double)myFile.getSize()));
+        simpleLogListProperty.setProcess(((double) curSize) / ((double) myFile.getSize()));
     }
 
     private void buildFile() throws IOException {
         realFile = new File(realPos);
-        infoFile = new File(infoFile+".info");
-        if(!realFile.exists())realFile.createNewFile();
-        if(!infoFile.exists())infoFile.createNewFile();
+        infoFile = new File(realPos + ".info");
+        if (!realFile.exists()) realFile.createNewFile();
         else{
+            curSize = realFile.length();
+        }
+        if (!infoFile.exists()) infoFile.createNewFile();
+        else {
             Scanner scan = new Scanner(realFile);
-            if(scan.hasNext())
+            if (scan.hasNext())
                 md5 = scan.next();
-            if(scan.hasNextLong())
+            if (scan.hasNextLong())
                 wholeSize = scan.nextLong();
         }
     }
@@ -57,27 +63,25 @@ public class DownLoader implements Runnable, Serializable {
         flag = true;
     }
 
-
-    @Override
-    public void run() {
-
+    @Deprecated
+    private void work(){
         BufferedOutputStream bufferedOutputStream = null;
         try {
-            bufferedOutputStream = new BufferedOutputStream(new FileOutputStream(realFile));
+            bufferedOutputStream = new BufferedOutputStream(new FileOutputStream(realFile, true));
         } catch (FileNotFoundException e) {
             e.printStackTrace();
             return;
         }
         byte[] buffer = new byte[PACKAGE_SIZE];
         int tmp = 0;
-        while(curSize < wholeSize && flag){
+        while (curSize < wholeSize && (!flag)) {
             try {
-                tmp = server.readByteOfFile(id,pos,name,buffer,0,PACKAGE_SIZE);
-                bufferedOutputStream.write(buffer,0,tmp);
+                tmp = server.readByteOfFile(id, pos, name, buffer, 0, PACKAGE_SIZE);
+                bufferedOutputStream.write(buffer, 0, tmp);
                 curSize += tmp;
 
-                if((double)curSize/wholeSize > simpleLogListProperty.getProcess() + 0.1){
-                    simpleLogListProperty.setProcess((double)curSize/wholeSize);
+                if ((double) curSize / wholeSize > simpleLogListProperty.getProcess() + 0.01) {
+                    simpleLogListProperty.setProcess((double) curSize / wholeSize);
                 }
 
             } catch (ClassNotFoundException | NoUserException | NoFileException | IOException | NoAccessException e) {
@@ -86,8 +90,55 @@ public class DownLoader implements Runnable, Serializable {
             }
 
         }
-
+        try {
+            bufferedOutputStream.close();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
+        if (flag) {
+            return;
+        }
         simpleLogListProperty.setProcess(1);
         infoFile.delete();
+    }
+
+    private void transfer() throws IOException {
+        BufferedOutputStream bufferedOutputStream = new BufferedOutputStream(new FileOutputStream(realFile,true));
+        byte[] buffer = new byte[PACKAGE_SIZE];
+
+        long pretime = MyDate.getNowTimeStamp();
+        long preData = curSize;
+
+        while(curSize < wholeSize && (!flag)){
+            int tmp = serverReader.read(buffer,0,PACKAGE_SIZE);
+            bufferedOutputStream.write(buffer,0,tmp);
+            curSize += tmp;
+            if ((double) curSize / wholeSize > simpleLogListProperty.getProcess() + 0.01) {
+                simpleLogListProperty.setProcess((double) curSize / wholeSize);
+            }
+            if(MyDate.getNowTimeStamp() - pretime >= 1){
+                long delta = MyDate.getNowTimeStamp() - pretime;
+                delta = (curSize - preData)/delta;
+                simpleLogListProperty.setSpeed(MyFile.transferSize(delta)+"/s");
+                pretime = MyDate.getNowTimeStamp();
+                preData = curSize;
+            }
+        }
+        if(curSize >= wholeSize){
+            infoFile.delete();
+            simpleLogListProperty.setProcess(1);
+        }
+        bufferedOutputStream.close();
+        serverReader.close();
+
+    }
+
+    @Override
+    public void run() {
+        try {
+            transfer();
+        } catch (IOException e) {
+            e.printStackTrace();
+        }
     }
 }
